React class App1 oluşturuldu.
$ npm install

When you have created the file, you can install the packages via npm install. This will create a node_modules folder, so we should also add that folder to .gitignore. If you don't have npm installed, now is a good time to google for it and find out how to install it on your OS.